add in system metrics to the deployment details dashboard
