+++ datmo/dashboard/templates/model_summary.html
@@ -47,7 +47,7 @@
 
                         <div class="multi-item-content-row snapshot-filters">
                             <div class="input-group pull-right">
-                                <a href="/model/{{ model.name }}/snapshots/.csv" target="_blank"><span class="btn"><i class="fa fa-download"></i></span></a>
+                                <a href="/model/{{ model.name }}/snapshots/.csv" target="_blank"><span class="btn blue"><i class="fa fa-download"></i></span></a>
                             </div>
                             <div class="">
                                 <p>
